finished all options classes
