--- menpowidgets/options_old/__init__.py
@@ -1,5 +1,0 @@
-from .base import (AnimationOptionsWidget, ChannelOptionsWidget,
-                   FeatureOptionsWidget, GraphOptionsWidget,
-                   LandmarkOptionsWidget, LinearModelParametersWidget,
-                   PatchOptionsWidget, RendererOptionsWidget,
-                   SaveFigureOptionsWidget, TextPrintWidget)
